requirements.txt
Flask == 1.0.2
Pillow == 7.2.0
numpy == 1.19.5
tensorflow == 2.4.1
transformers == 4.18.0

